chore(string): organize structure.
