+++ src/string/lib/index.ts
@@ -1,0 +1,1 @@
+export { StringState } from './string-state.abstract';
